requirements.txt
# ============================================
# Intelligent Scheduler - Backend Requirements
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI & AI
openai==1.10.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
celery==5.3.6

# LangChain & Agents (for future agents)
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20

# Google APIs (for calendar integration)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
email-validator==2.3.0
